wsi_sr/requirements.txt
torch>=2.0
torchvision>=0.15
pillow
numpy
scikit-image
lpips
onnx
onnxruntime
